frontend/requirements.txt
streamlit>=1.36
requests
pillow
